requirements.txt
accelerate==1.2.1
torch==2.5.1
pandas==2.2.3
transformers==4.48.0
tokenizers==0.21.0
datasets==3.2.0
matplotlib==3.10.0
torchmetrics==1.6.1
scikit-learn==1.6.1
tqdm==4.67.1
pyzmq==26.2.0
omegaconf==2.3.0
pyyaml==6.0.2
smact==3.0.1
matminer==0.9.3
loralib==0.1.2
wandb==0.19.2
codecarbon==2.8.2
psutil==6.1.1
pymatgen==2025.1.9
numpy==1.26.4
flash-attention==1.0.0
accelerate==1.2.1
commentjson==0.9.0
deepspeed==0.16.3
optuna==4.2.1
ipywidgets==8.1.6
jarvis-tools==2024.10.30
trl==0.16.1
brokenaxes==0.6.2
mp-api==0.45.5
seaborn==0.13.2
mace-torch==0.3.13
crystal_toolkit==2025.7.31
# pip install git+https://github.com/lematerial/material-hasher.git